requirements.txt
pandas
Pillow
scikit-learn
tensorflow
numpy
matplotlib
streamlit
Datetime
pickle5
